# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/documents/{id}/draft" : {
      "get" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "get_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DraftView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "put_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PutDraftRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DraftView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "delete_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/documents/{id}/draft" : {
      "get" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "get_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DraftView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "put_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PutDraftRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DraftView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-draft-controller" ],
        "operationId" : "delete_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/documents/{id}" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "get_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-controller" ],
        "operationId" : "update_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-controller" ],
        "operationId" : "softDelete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "get_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-controller" ],
        "operationId" : "update_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-controller" ],
        "operationId" : "softDelete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approval-role-assignments" : {
      "get" : {
        "tags" : [ "approval-role-assignment-controller" ],
        "operationId" : "list_9",
        "parameters" : [ {
          "name" : "roleId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "spaceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AssignmentView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "approval-role-assignment-controller" ],
        "operationId" : "assign",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssignmentView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tags/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportTag",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/tags/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportTag_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/folders/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportFolder",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/folders/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportFolder_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/export" : {
      "get" : {
        "tags" : [ "export-controller" ],
        "operationId" : "exportDocument_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approvals/{requestId}" : {
      "get" : {
        "tags" : [ "approval-controller" ],
        "operationId" : "get_18",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApprovalDetailView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trash/documents/{documentId}" : {
      "delete" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "softDeleteDocument",
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/approval-role-assignments/{id}" : {
      "delete" : {
        "tags" : [ "approval-role-assignment-controller" ],
        "operationId" : "unassign",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DocumentPermissions" : {
        "type" : "object",
        "properties" : {
          "canEdit" : {
            "type" : "boolean"
          },
          "canPublish" : {
            "type" : "boolean"
          },
          "canManageAccess" : {
            "type" : "boolean"
          },
          "canComment" : {
            "type" : "boolean"
          },
          "canManageAttestations" : {
            "type" : "boolean"
          }
        }
      },
      "DocumentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "spaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "folderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "status" : {
            "type" : "string"
          },
          "currentVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reliabilityScore" : {
            "type" : "number"
          },
          "reliabilityComputedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stale" : {
            "type" : "boolean"
          },
          "contentModifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stalenessThresholdDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "visibility" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "templateId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "templateVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "updatedBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "owner" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TagRef"
            }
          },
          "permissions" : {
            "$ref" : "#/components/schemas/DocumentPermissions"
          }
        }
      },
      "PersonRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "initials" : {
            "type" : "string"
          }
        }
      },
      "TagRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "governed" : {
            "type" : "boolean"
          }
        }
      },
...
      "PutDraftRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "baseVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "DraftView" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "baseVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "UpdateDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "minLength" : 1
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "changeSummary" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "expectedVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "body", "title" ]
      },
...
      "Body" : {
        "type" : "object",
        "properties" : {
          "roleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "subjectType" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "scopeType" : {
            "type" : "string",
            "minLength" : 1
          },
          "scopeRef" : {
            "type" : "string"
          }
        },
        "required" : [ "roleId", "scopeType", "subjectId", "subjectType" ]
      },
      "AssignmentView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "roleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "roleName" : {
            "type" : "string"
          },
          "subjectType" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "scopeType" : {
            "type" : "string"
          },
          "scopeRef" : {
            "type" : "string"
          },
          "grantedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "grantedAt" : {
            "type" : "string"
          }
        }
      },
...
      "ImpactedLink" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          }
        }
      },
...
      "ApprovalDetailView" : {
        "type" : "object",
        "properties" : {
          "approvalRequestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "documentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "documentTitle" : {
            "type" : "string"
          },
          "temporalWorkflowId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "currentStepOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slaDeadlineAt" : {
            "type" : "string"
          },
          "submittedVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "baselineVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string"
          },
          "requestedByDisplayName" : {
            "type" : "string"
          },
          "requestedByInitials" : {
            "type" : "string"
          },
          "impactedLinks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImpactedLink"
            }
          },
          "hiddenImpactedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "canDecide" : {
            "type" : "boolean"
          },
          "cannotDecideReason" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}